Updating the database connection string to use keyring
We now store the db password in the system's keyring. That way it's stored in an encrypted location only accessible by the logged in user. We don't have to worry about it being committed to the repository by mistake and we can provide a configuration that should work for most default installations.