Add a generic JSON consumer
Requires confluent, curses, and a few fairly generic classes.

Takes up to three arguments on the command line; the name of the kafka topic to take input from, the addressof the Kafka server to readfrom, and the group ID to use when reading. (If any of these is not provided, default values will be used instead, which is likely to be harmless but probably not what you're looking for). To exit, give it a keypress - any keypress.

Graphs are not guaranteed to be meaningful, However, it should help to test almost any producer.

Now rebased against the master branch.